requirements.txt
# === LangGraph + LangChain ===
langgraph>=0.2.28
langchain>=0.3.14
langchain-groq>=0.2.4
langchain-core>=0.3.28
langchain-community>=0.3.14

# === LLM Provider ===
groq>=0.12.0

# === Vector Store (RAG) ===
chromadb>=0.5.23
sentence-transformers>=3.3.1

# === WhatsApp ===
twilio>=9.4.0

# === Data Validation ===
pydantic>=2.10.0
pydantic-settings>=2.7.0

# === Async ===
httpx>=0.28.0
aiofiles>=24.1.0

# === Utilities ===
python-dotenv>=1.0.1
tenacity>=9.0.0

# === PDF Extraction (Form 16 & CAS) ===
pdfplumber>=0.11.0
casparser>=0.6.0

# === Testing ===
pytest>=8.3.4
pytest-asyncio>=0.24.0
